requirements.txt
azure-cognitiveservices-speech==1.31.0
python-dotenv==1.0.0
flask==3.0.0
flask-socketio==5.3.6 